# src/main/resources/Swagger_feedback.yaml
---
openapi: 3.1.0
info:
  description: Ecohub API Documentation
  title: Feedback Form
  version: '0.0.1'
servers:
- description: ecohub server
  url: http://localhost:8080
paths:
  "/admin/messages":
    get:
      description: Получение списка сообщений от пользователей.
      tags: 
        - Administrator actions
      operationId: getMessages
      responses: 
        '200':
          description: получение списка сообщений
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Messages'
        '404':
          description: сообщения не найдены
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessageError'
  "/admin/messages/{id}":
    get:
      description: получение сообщения с заданным id от пользователя
      tags: 
        - Administrator actions
      operationId: getMessage
      parameters: 
      - name: id 
        in: path
        required: true
        schema:
          type: integer
        example: 1
      responses: 
        '200':
          description: получение сообщения
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: сообщение не найдено
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessageError'
    post: 
      description: отправка ответа пользователю
      tags: 
        - Administrator actions
      operationId: sentMessageByAdmin
      requestBody: 
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses: 
        '200':
          description: сообщение отправлено
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: сообщение не отправлено
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/MessageNotSentError'
  "/feedback":
    post:
      description: отправка сообщения пользователем
      tags: 
        - User actions
      operationId: sentMessageByUser
      requestBody: 
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Message'

components:
  schemas: 
    Message:
      type: object
      required: 
        - feedback_id
      properties:
        feedback_id:
          type: integer
          example: 12
        email:
          type: string
          example: example@gmail.com
        message_topic:
          type: string
          description: краткое описание темы обращения/вопроса, для удобной навигации
          example: неверная информация на сайте
        message_content:
          type: string
          description: тело вопроса который задает пользователь
          example: Добрый день! Проверьте, пожалуйста, контактные данные по точке переработки по улице Слободская 59, не получается дозвониться по указанному номеру.
        message_time:
          type: string
          description: время, когда вопрос пользователя был отправлен
          format: date-time
          example: 03-07-2023 20:17:11
        admin_id:
          type: integer
          description: id администратора, который ответил пользователю
          example: 1
        response_content:
          type: string
          description: текст ответа от администратора
          example: Добрый день! Спасибо за Ваше обращение, информация по пункту переработки, находящемуся по адресу ул.Слободская 59 успешно обновлена. С уважением, Алеся.
        response_time:
          type: string
          description: время отправки ответа пользователю на его вопрос
          format: date-time
          example: 04-07-2023 09:00:16
        response_status:
          type: string
          description: когда пользователь отправляет сообщение, ему присваивается статус "open", после того как сообщение открывается и просматривается администратором его статус меняется на "in_progress", и когда администратор отправляет ответ, статус фидбэка меняется на " closed"
          enum: 
            - closed
            - open
            - in_progress
            
    Messages:
      type: array
      items:
        $ref: '#/components/schemas/Message'
    GetMessageError:
      type: object
      required: 
        - code
        - message
      properties: 
        code:
          type: integer
          example: 404
        message:
          type: string
          description: описание ошибки
          example: NOT_FOUND
    MessageNotSentError:
      type: object
      required: 
        - code
        - message
      properties: 
        code:
          type: integer
          example: 420
        message:
          type: string
          description: описание ошибки
          example: METHOD_FAILURE